requirements.txt
# AWS CDK Dependencies
aws-cdk-lib>=2.100.0
constructs>=10.0.0

# Python Dependencies
boto3>=1.34.0
botocore>=1.34.0

# Development Dependencies
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0

# Testing Dependencies
pytest-mock>=3.11.0
moto>=4.2.0

# Documentation Dependencies
sphinx>=7.0.0
sphinx-rtd-theme>=1.3.0
